Cleaned up the demo
Formatted the HTML, added the Roboto and Roboto Mono fonts, and
switched to the Atom One Light highlight.js theme, which looks
a lot better than the default.

I've also increased the line height of the text, which looks so
much better in my opinion, and I've added a link to the GitHub
repository in the header so folks can get back to the source
quickly.